// src/generated/accounts/Remote.ts
export type RemoteArgs = {
  address: number[] /* size: 32 */
  bump: number
}

export const remoteDiscriminator = [142, 141, 61, 250, 205, 173, 146, 34]
/**
 * Holds the data for the {@link Remote} Account and provides de/serialization
 * functionality for that data
 *
 * @category Accounts
 * @category generated
 */
export class Remote implements RemoteArgs {
  private constructor(
    readonly address: number[] /* size: 32 */,
    readonly bump: number
  ) {}

  /**
   * Creates a {@link Remote} instance from the provided args.
   */
  static fromArgs(args: RemoteArgs) {
    return new Remote(args.address, args.bump)
  }

  /**
   * Deserializes the {@link Remote} from the data of the provided {@link web3.AccountInfo}.
   * @returns a tuple of the account data and the offset up to which the buffer was read to obtain it.
   */
  static fromAccountInfo(
    accountInfo: web3.AccountInfo<Buffer>,
    offset = 0
  ): [Remote, number] {
    return Remote.deserialize(accountInfo.data, offset)
  }

  /**
   * Retrieves the account info from the provided address and deserializes
   * the {@link Remote} from its data.
   *
   * @throws Error if no account info is found at the address or if deserialization fails
   */
  static async fromAccountAddress(
    connection: web3.Connection,
    address: web3.PublicKey,
    commitmentOrConfig?: web3.Commitment | web3.GetAccountInfoConfig
  ): Promise<Remote> {
    const accountInfo = await connection.getAccountInfo(
      address,
      commitmentOrConfig
    )
    if (accountInfo == null) {
      throw new Error(`Unable to find Remote account at ${address}`)
    }
    return Remote.fromAccountInfo(accountInfo, 0)[0]
  }

  /**
   * Provides a {@link web3.Connection.getProgramAccounts} config builder,
   * to fetch accounts matching filters that can be specified via that builder.
   *
   * @param programId - the program that owns the accounts we are filtering
   */
  static gpaBuilder(
    programId: web3.PublicKey = new web3.PublicKey(
      '4JSbnmWnAmRfaCWeZwFHWJrakYgrqtkXRqZ6MN4o9KiE'
    )
  ) {
    return beetSolana.GpaBuilder.fromStruct(programId, remoteBeet)
  }

  /**
   * Deserializes the {@link Remote} from the provided data Buffer.
   * @returns a tuple of the account data and the offset up to which the buffer was read to obtain it.
   */
  static deserialize(buf: Buffer, offset = 0): [Remote, number] {
    return remoteBeet.deserialize(buf, offset)
  }

  /**
   * Serializes the {@link Remote} into a Buffer.
   * @returns a tuple of the created Buffer and the offset up to which the buffer was written to store it.
   */
  serialize(): [Buffer, number] {
    return remoteBeet.serialize({
      accountDiscriminator: remoteDiscriminator,
      ...this,
    })
  }

  /**
   * Returns the byteSize of a {@link Buffer} holding the serialized data of
   * {@link Remote}
   */
  static get byteSize() {
    return remoteBeet.byteSize
  }

  /**
   * Fetches the minimum balance needed to exempt an account holding
   * {@link Remote} data from rent
   *
   * @param connection used to retrieve the rent exemption information
   */
  static async getMinimumBalanceForRentExemption(
    connection: web3.Connection,
    commitment?: web3.Commitment
  ): Promise<number> {
    return connection.getMinimumBalanceForRentExemption(
      Remote.byteSize,
      commitment
    )
  }

  /**
   * Determines if the provided {@link Buffer} has the correct byte size to
   * hold {@link Remote} data.
   */
  static hasCorrectByteSize(buf: Buffer, offset = 0) {
    return buf.byteLength - offset === Remote.byteSize
  }

  /**
   * Returns a readable version of {@link Remote} properties
   * and can be used to convert to JSON and/or logging
   */
  pretty() {
    return {
      address: this.address,
      bump: this.bump,
    }
  }
}

/**
 * @category Accounts
 * @category generated
 */
export const remoteBeet = new beet.BeetStruct<
  Remote,
  RemoteArgs & {
    accountDiscriminator: number[] /* size: 8 */
  }
>(
  [
    ['accountDiscriminator', beet.uniformFixedSizeArray(beet.u8, 8)],
    ['address', beet.uniformFixedSizeArray(beet.u8, 32)],
    ['bump', beet.u8],
  ],
  Remote.fromArgs,
  'Remote'
)